requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
contextlib2==21.6.0
cryptography==46.0.3
Django==5.2.5
django-debug-toolbar==6.0.0
django-jazzmin==3.0.1
django-ninja==1.4.5
django-ninja-extra==0.30.2
django-ninja-jwt==5.4.0
idna==3.11
injector==0.22.0
pillow==11.3.0
psycopg2==2.9.11
pycparser==2.23
pydantic==2.12.2
pydantic_core==2.41.4
PyJWT==2.10.1
pyTelegramBotAPI==4.29.1
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.3
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
